pyproject.toml
[project]
name = "linkedin-blog-sync"
version = "0.1.0"
description = "Sync blog posts from eve.gd to LinkedIn"
readme = "SETUP.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "markdown>=3.5.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
]

[project.scripts]
linkedin-sync = "sync:cli"
linkedin-oauth = "oauth_helper:get_access_token"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=4.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
